+++ src/main/java/net/app/savable/controller/VerificationController.java
@@ -6,8 +6,8 @@ import lombok.extern.slf4j.Slf4j;
 import net.app.savable.domain.challenge.ParticipationChallenge;
 import net.app.savable.domain.challenge.VerificationState;
 import net.app.savable.domain.challenge.dto.VerificationRequestDto;
-import net.app.savable.global.common.ApiResponse;
-import net.app.savable.global.common.ErrorCode;
+import net.app.savable.global.error.ApiResponse;
+import net.app.savable.global.error.exception.ErrorCode;
 import net.app.savable.global.common.S3UploadService;
 import net.app.savable.service.ParticipationChallengeService;
 import net.app.savable.service.VerificationService;
